✨ Expose the embeddable color scheme editor and an external customize hook on the theme toggle.
Extract HkColorSchemeEditor from HkColorSchemeDialog so downstream apps can host the token/group editing form inside their own surface (tabs, drawers) while the dialog keeps its exact public API as a thin modal wrapper. Add the externalCustomize prop and open-customize emit to HkThemeToggle so a host app can own the customization entry point, and label the entry with the new hikari::theme.customize key in all eleven locales.
