+++ packages/vue/src/i18n/locales/zh-Hans/platform.json
@@ -22,6 +22,7 @@
     "hikari::theme.modeAuto": "自动",
     "hikari::theme.themes": "配色主题",
     "hikari::theme.editScheme": "编辑配色方案…",
+    "hikari::theme.customize": "自定义",
     "hikari::theme.themeName": "主题名称",
     "hikari::theme.customThemeName": "自定义主题",
     "hikari::theme.save": "保存",
